stars reviews

+++ db/migrate/20230606122843_add_reviewer_to_review.rb
@@ -1,5 +1,5 @@
 class AddReviewerToReview < ActiveRecord::Migration[6.1]
   def change
-    add_column :reviews, :reviewer, :integer
+    add_column :reviews, :reviewer, :bigint
   end
 end
